// src/generated/model/detailedMapping.ts
/**
 * ModMappings API
 * This is the api for ModMappings. It is currently in development and in an alpha stage.
 *
 * The version of the OpenAPI document: 0.0.0-Dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Mappable } from './mappable';
import { SimpleVersionedMappable } from './simpleVersionedMappable';
import { Mapping } from './mapping';


/**
 * Represents a single piece of the sourcecode whose name can be remapped in a given version of the game, including its metadata and core information.
 */
export interface DetailedMapping { 
    mappable?: Mappable;
    versionedMappable?: SimpleVersionedMappable;
    mappingDTO?: Mapping;
}