// src/generated/model/mappableType.ts
/**
 * ModMappings API
 * This is the api for ModMappings. It is currently in development and in an alpha stage.
 *
 * The version of the OpenAPI document: 0.0.0-Dev
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Indicates which types (as in kind of pieces of the source code) can be remapped.
 */
export type MappableType = 'CLASS' | 'METHOD' | 'FIELD' | 'PARAMETER';

export const MappableType = {
    CLASS: 'CLASS' as MappableType,
    METHOD: 'METHOD' as MappableType,
    FIELD: 'FIELD' as MappableType,
    PARAMETER: 'PARAMETER' as MappableType
};